cleanup current keymaps
