+++ packages/main/src/modules/KeyHandler/defaultKeyMaps/gw2_minstrel.ts
@@ -1,0 +1,19 @@
+import { KeyMap } from "../types.js";
+
+export const keyMap: KeyMap = {
+  name: 'GW2 - The Minstrel (C)',
+  autoOctaveSwap: false,
+  notes: {
+    C4: { key: '1' },
+    D4: { key: '2' },
+    E4: { key: '3' },
+    F4: { key: '4' },
+    G4: { key: '5' },
+    A4: { key: '6' },
+    B4: { key: '7' },
+    C5: { key: '8' },
+
+    'C#4': { key: '9' },
+    'D#4': { key: '0' },
+  },
+}
